feat(header): extract Hero, HeaderBar and Logo components for better shadowing support
