+++ examples/full/src/gatsby-plugin-theme-ui/components.js
@@ -8,7 +8,7 @@ import Column from '@gatsby-mdx-suite/mdx-layout/column'
 import Section from '@gatsby-mdx-suite/mdx-layout/section'
 
 // Project specific components
-import Header from 'gatsby-theme-mdx-suite-base/src/components/layout/header'
+import Header from 'gatsby-theme-mdx-suite-base/src/components/header/header'
 import CustomComponent from '../components/custom-component'
 
 export default {
