+++ themes/base/src/components/header/logo.js
@@ -1,0 +1,6 @@
+import Logo from '../../assets/logo.svg'
+
+// Use this extra file for dynamic logos or when you can not use SVGs.
+// Otherwise just shadow the svg from above.
+
+export default Logo
