+++ themes/blog/src/templates/blog-post.js
@@ -8,7 +8,7 @@ import MdxSuiteContext from '@gatsby-mdx-suite/contexts/mdx-suite'
 import mergeContextData from '@gatsby-mdx-suite/helpers/data/merge-context-data'
 
 import Layout from 'gatsby-theme-mdx-suite-base/src/components/layout/layout'
-import Header from 'gatsby-theme-mdx-suite-base/src/components/layout/header'
+import Header from 'gatsby-theme-mdx-suite-base/src/components/header/header'
 import Seo from 'gatsby-theme-mdx-suite-base/src/components/layout/seo'
 
 function BlogPostTemplate({ data, pageContext }) {
